refactor: move Video to another comp
